feat: support extraction of MJPEG video frames
This commit adds support for extracting MJPEG files out of
.vraw-files.

The generated files are playable in `ffplay`, but not VLC. This might
be because MJPEG is not a standardized format and therefore not
documented where it is easily found.

To do something interesting with the extracting MJPEG files it needs
to be converted with for instance `ffmpeg`.